Move contract event parsing to its own package
